feat(ui): home scorecard + slim dashboard + Govern 3-group nav
